pyproject.toml
[project]
name = "gogo_keyboard"
version = "1.0.2"
description = "Press keyboard 🦍 Get key 🦍 Python Asyncio for Robotics"
readme = "README.md"
license = "MIT"
authors = [
  { name = "Elian NEPPEL", email = "elian.dev@posteo.com" }
]
requires-python = ">=3.10"
keywords = ["asyncio", "keyboard", "ros2", "zenoh"]

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Environment :: Console",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: MacOS",
  "Topic :: Scientific/Engineering",
  "Framework :: AsyncIO",
]

dependencies = [
    "asyncio_for_robotics>=1.3.1",
    "pysdl2>=0.9.0",
]

[build-system]
requires = ["setuptools >= 58.2.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"gogo_keyboard" = [
  "py.typed",
  "icons/gogo.png",
  "icons/gogo_happy.png",
  "icons/gogo_happy2.png",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "gogo_keyboard",
]



[project.optional-dependencies]
dll = ["pysdl2-dll"]

[project.urls]
Homepage = "https://github.com/2lian/gogo_keyboard"
Repository = "https://github.com/2lian/gogo_keyboard"
Issues = "https://github.com/2lian/gogo_keyboard/issues"
